+++ src/main/java/Spring/MindStone/domain/friend/MessageList.java
@@ -1,12 +1,11 @@
-package Spring.MindStone.domain;
+package Spring.MindStone.domain.friend;
 
 import Spring.MindStone.domain.enums.MessageStatus;
+import Spring.MindStone.domain.member.MemberInfo;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
