+++ src/main/java/Spring/MindStone/domain/habit/Habit.java
@@ -1,13 +1,11 @@
-package Spring.MindStone.domain;
+package Spring.MindStone.domain.habit;
+import Spring.MindStone.domain.member.MemberInfo;
 import Spring.MindStone.domain.common.BaseEntity;
-import Spring.MindStone.domain.enums.MessageStatus;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 @Entity
 @Getter
 @Builder
