+++ src/main/java/Spring/MindStone/domain/member/MemberProfile.java
@@ -1,13 +1,6 @@
-package Spring.MindStone.domain;
-import Spring.MindStone.domain.common.BaseEntity;
-import Spring.MindStone.domain.enums.MessageStatus;
+package Spring.MindStone.domain.member;
 import jakarta.persistence.*;
 import lombok.*;
-import org.hibernate.annotations.DynamicInsert;
-import org.hibernate.annotations.DynamicUpdate;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
